+++ models/user_post.go
@@ -91,7 +91,7 @@ func (m *UserPost) FindById(id int) (userPost UserPost, err error) {
 	return
 }
 
-func (m *UserPost) FindByMap(offset, limit int, dataMap map[string]interface{},orderBy string) (res []UserPost, total int, err error) {
+func (m *UserPost) FindByMap(offset, limit int64, dataMap map[string]interface{},orderBy string) (res []UserPost, total int64, err error) {
 	query := Db
 	if status,isExist:=dataMap["status"].(int);isExist{
 		query = query.Where("status = ?", status)
